refactor: extract saveAllCollectionChanges to Redux actions for better testability and reuse
